utilities/shopping_list_api/pyproject.toml
[project]
name = "shopping-list-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "google-auth>=2.40.3",
    "google-genai>=1.28.0",
    "openai>=1.98.0",
    "uvicorn>=0.35.0",
]
